Approx 30x speed up on VarPasswordPair clone()
``` VarPasswordPair ``` is forced to use ``` Secret.fromString() / Secret.toString() ``` as there's no public ``` Secret ``` constructor. This results in an attempted decrypt - which fails throwing a ``` GeneralSecurityException ```. Unfortunately that's not very performant and this is exacerbated on large Jenkins installations with several global Masked Passwords - particularly so when you're using the BuildPipeline plugin and the parameterized trigger plugin.