Require Jenkins 2.361.4 or newer (#30)
* Require Jenkins 2.361.4 or newer

Updates to the most recent parent pom and updates to require Java 11
and Jenkins 2.361.4 or newer.

https://stats.jenkins.io/pluginversions/mask-passwords.html shows that
60% of the installations of the most recent release of the plugin were
already running at least Jenkins 2.361.1 as of the early January 2023
run of that installation count report.  Since a majority of users of
the current version are already on Jenkins 2.361.1 or newer, let's
modernize the plugin to require Java 11 and Jenkins 2.361.4 or newer.

https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/
recommends either 2.361.4 or 2.346.3 as minimum required version.  By
choosing 2.361.4, the plugin moves from Java 8 to Java 11.

* Test Java 17 in CI and use container agents

Modernize the Jenkinsfile so that it tests both Java 11 and Java 17
and so that it uses container agents that are faster to start and
easier to maintain.